% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addCite.R
\name{addCite}
\alias{addCite}
\title{include citation for package in markdown}
\usage{
addCite(bib, packages, type = "T")
}
\arguments{
\item{packages}{}
}
\description{
include citation for package in markdown
}

