% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listcitekeys.R
\name{listcitekeys}
\alias{listcitekeys}
\title{list the citation keys for a package}
\usage{
listcitekeys(packages, title = F)
}
\arguments{
\item{the}{package name}
}
\description{
list the citation keys for a package
}

