% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logTime.R
\name{logTime}
\alias{logTime}
\title{Load data from OfficeTime export into my time recording DB
Takes a file name and squirts the info into the time recording database. It will check if
data already exists for the dates and stop rather than write duplicate entries.}
\usage{
logTime(filename = NULL, projDir = "/Users/pschofield/Projects/time_sheet/",
  dbName = "timeSheet.db")
}
\arguments{
\item{filename}{}
}
\description{
Load data from OfficeTime export into my time recording DB
Takes a file name and squirts the info into the time recording database. It will check if
data already exists for the dates and stop rather than write duplicate entries.
}

