% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newProject.R
\name{newProject}
\alias{newProject}
\title{create a new project}
\usage{
newProject(name = NULL, root = "/Users/pschofield/git_tree/", noidx = F,
  noDir = F, noWWW = F)
}
\arguments{
\item{noWWW}{}
}
\description{
create a new project
}

