% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcur.R
\name{rcur}
\alias{rcur}
\title{render the current document}
\usage{
rcur(fileName = .curFile, projDir = .curProj, inDir = NULL,
  outDir = "Notes/Projects", rootDir = "/Volumes/pschofield",
  codeDir = "/Users/pschofield/Code/bb/CRUK")
}
\arguments{
\item{pname}{project name}

\item{fname}{file name}

\item{tname}{tree name}

\item{rname}{root of tree}

\item{oflag}{open a file}

\item{rootdir}{root}
}
\description{
render the current document
}

