% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/subScript.R
\name{subScript}
\alias{subScript}
\title{subScript will submit a script to the cluster}
\usage{
subScript(scriptstub = "ssh", script = c("#!/bin/bash", "hostname"),
  tmpdir = "/homes/pschofield/tmp/", scriptext = ".sh", logdir = "",
  cores = 8)
}
\arguments{
\item{scriptstub}{the stub for the temporary file name}

\item{script}{the content of the script as a vector of strings}

\item{tmpdir}{location for the temporary file}

\item{scriptext}{extention for the temporary file}

\item{logdir}{location for the batch job logs}

\item{cores}{number of cores}
}
\description{
calls subJob to submit a script to the cluster as a temporary file it relies on the temporary
directory where the temporary file will be written to being mounted to the local machine.
I could get round this by writing it locally and then copying it with scp but at the moment
this is not worth the effort.
}

