% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genBibtex.R
\name{genBibtex}
\alias{genBibtex}
\title{genbibtex}
\usage{
genBibtex(packages, db = F,
  outfile = "/Users/pschofield/git_tree/biblio/library.bib")
}
\arguments{
\item{packages}{package names to add citations from}

\item{outfile}{file of bibtex file to add citations to}
}
\description{
genbibtex
}

