% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genBibtex.R
\name{genBibtex}
\alias{genBibtex}
\title{genbibtex}
\usage{
genBibtex(packages, db = F,
  outfile = "/Users/pschofield/git_tree/biblio/library.bib")
}
\arguments{
\item{packages}{}

\item{outfile}{}
}
\description{
genbibtex
}

