% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish}
\alias{publish}
\title{Publish all relevant files from the local directory to my public_html tree}
\usage{
publish(filePath, note = T, outStub = NULL, delete = F)
}
\arguments{
\item{rstudio}{}
}
\description{
Publish all relevant files from the local directory to my public_html tree
}

