% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readENCODE.R
\name{readENCODE}
\alias{readENCODE}
\title{readENCODE read broadPeak and narrowPeak formats and return a GRanges object}
\usage{
readENCODE(fn)
}
\arguments{
\item{filename}{}
}
\description{
readENCODE read broadPeak and narrowPeak formats and return a GRanges object
}

